registration-app/app/requirements.txt
flask==3.0.3
psycopg2-binary==2.9.10
gunicorn==23.0.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-instrumentation-psycopg2==0.48b0
